pyproject.toml
[project]
name = "we0-index"
version = "0.1.0"
description = "We0 Index"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "boto3>=1.36.18",
    "chromadb>=0.6.3",
    "fastapi>=0.115.7",
    "gitpython>=3.1.42",
    "greenlet>=3.1.1",
    "loguru>=0.7.3",
    "minio>=7.2.15",
    "openai",
    "opendal>=0.45.15",
    "psycopg[binary,pool]>=3.2.4",
    "pydantic-settings>=2.7.1",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.2",
    "qdrant-client>=1.13.2",
    "sqlalchemy>=2.0.38",
    "tiktoken>=0.8.0",
    "torch>=2.6.0",
    "transformers>=4.49.0",
    "tree-sitter>=0.24.0",
    "tree-sitter-css>=0.23.2",
    "tree-sitter-go>=0.23.4",
    "tree-sitter-java>=0.23.5",
    "tree-sitter-javascript>=0.23.1",
    "tree-sitter-python>=0.23.6",
    "tree-sitter-typescript>=0.23.2",
    "uvicorn>=0.34.0",
]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple"
default = true
